Refactor chart data handling and reorganize Jotai stores for improved structure #2
